+++ internal/generator/marshaling/doc.go
@@ -1,0 +1,3 @@
+// Package marshaler generates binary marshaler/unmarshaler methods for protobuf
+// message types.
+package marshaling