+++ CHANGELOG.md
@@ -7,8 +7,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 <!-- references -->
 
-[Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
-[Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/
+[semantic versioning]: https://semver.org/spec/v2.0.0.html
+[bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
+
+## [Unreleased]
+
+### Added
+
+- Added `MarshalBinary()` and `UnmarshalBinary()` methods such that all message
+  types implement the standard library's `encoding.BinaryMarshaler` and
+  `encoding.BinaryUnmarshaler` interfaces.
 
 ## [0.3.1] - 2024-07-16